requirements.txt
streamlit>=1.30.0
openai>=1.3.0
fastapi>=0.104.0
uvicorn>=0.23.0
pdfplumber>=0.10.0
python-docx>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0 
streamlit
requests
python-dotenv
python-multipart
fastapi
uvicorn
python-dotenv
openai
pdfplumber
python-docx
pytesseract
pdf2image
python-multipart
Pillow